requirements.txt
roboflow
opencv-python-headless
pillow
python-dotenv
flask
flask-cors
werkzeug
requests
groq
matplotlib
flask-login
flask-sqlalchemy
numpy
scikit-learn
xgboost
lime
shap
scikit-image
reportlab